+++ gdtoolkit/formatter/__init__.py
@@ -12,12 +12,14 @@ from .safety_checks import (  # noqa: F401
 )
 
 
+# pylint: disable-next=too-many-arguments
 def check_formatting_safety(
     given_code: str,
     formatted_code: str,
     max_line_length: int,
     given_code_parse_tree: Optional[Tree] = None,
     given_code_comment_parse_tree: Optional[Tree] = None,
+    spaces_for_indent: Optional[int] = None,
 ) -> None:
     if given_code == formatted_code:
         return
@@ -40,4 +42,5 @@ def check_formatting_safety(
         max_line_length,
         parse_tree=formatted_code_parse_tree,
         comment_parse_tree=formatted_code_comment_parse_tree,
+        spaces_for_indent=spaces_for_indent,
     )