+++ gdtoolkit/formatter/constants.py
@@ -1,8 +1,7 @@
 from types import MappingProxyType
 
 
-INDENT_STRING = "\t"
-INDENT_SIZE = 4
+TAB_INDENT_SIZE = 4
 INLINE_COMMENT_OFFSET = 2
 
 DEFAULT_SURROUNDING_EMPTY_LINES_TABLE = MappingProxyType(