+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - Default regex for names of constants now allows underscore as a prefix to denote private contants (#223)
  - Fixed parsing of files without newline at the end of file ending with comment (#237)
  - Fixed linter support for docstrings (#233)
+ - Fixed linter support for trailing comma in function's args list (#206)
 
 ## [4.1.0] 2023-07-06